Integrating AbstractMCMC and fixing predictions (#93)
* Started adding AbstractMCMC sampling

* Removed part in training.jl

* Fixed the predictions

* Fixing docs

* Update bayesiansvm.jl

* Update logistic.jl

* Update gibbssampling.jl

* Update make.jl

* Correct formatting and tests

* Fix MCVI (maybe?)

* Improve the docs

* Fix some mroe tests

* Fixed EVEN more bugs

+++ src/likelihood/logisticsoftmax.jl
@@ -192,9 +192,9 @@ end
 function grad_samples(
     model::AbstractGP{T,<:LogisticSoftMaxLikelihood,<:NumericalVI},
     samples::AbstractMatrix{T},
-    index::Integer,
+    index::Int,
 ) where {T}
-    class = model.likelihood.y_class[index]::Int64
+    class = model.likelihood.y_class[index]::Int
     grad_μ = zeros(T, nLatent(model))
     grad_Σ = zeros(T, nLatent(model))
     g_μ = similar(grad_μ)
